frontend/requirements.txt
flask
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-http
